+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/repository/RoleRepository.java
@@ -35,7 +35,9 @@ public interface RoleRepository extends PagingAndSortingRepository<Role, Long> {
   @Query("SELECT r.id from Role r where r.roleName like CONCAT('Master+', ?1) "
       + "OR r.roleName like CONCAT('ModifyNamespace+', ?1, '+%') "
       + "OR r.roleName like CONCAT('ReleaseNamespace+', ?1, '+%')  "
-      + "OR r.roleName like CONCAT('ManageAppMaster+', ?1)")
+      + "OR r.roleName like CONCAT('ManageAppMaster+', ?1) "
+      + "OR r.roleName like CONCAT('ModifyNamespacesInCluster+', ?1, '+%')"
+      + "OR r.roleName like CONCAT('ReleaseNamespacesInCluster+', ?1, '+%')")
   List<Long> findRoleIdsByAppId(String appId);
 
   @Query("SELECT r.id from Role r where r.roleName like CONCAT('ModifyNamespace+', ?1, '+', ?2) "
@@ -47,4 +49,8 @@ public interface RoleRepository extends PagingAndSortingRepository<Role, Long> {
   @Modifying
   @Query("UPDATE Role SET IsDeleted = true, DeletedAt = ROUND(UNIX_TIMESTAMP(NOW(4))*1000), DataChange_LastModifiedBy = ?2 WHERE Id in ?1 and IsDeleted = false")
   Integer batchDelete(List<Long> roleIds, String operator);
+
+  @Query("SELECT r.id from Role r where r.roleName = CONCAT('ModifyNamespacesInCluster+', ?1, '+', ?2, '+', ?3) "
+      + "OR r.roleName = CONCAT('ReleaseNamespacesInCluster+', ?1, '+', ?2, '+', ?3)")
+  List<Long> findRoleIdsByCluster(String appId, String env, String clusterName);
 }